backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["christianmasa <christian@linkernetworks.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django-rest-framework = "^0.1.0"
django-cors-headers = "^3.13.0"
djangorestframework = "^3.13.1"
drf-spectacular = "^0.22.1"
opencv-python-headless = "^4.6.0"
scikit-image = "^0.19.3"
psycopg2 = "^2.9.3"
django-extensions = "^3.2.0"
django-celery-beat = "^2.3.0"
redis = "^4.3.4"
gunicorn = "^20.1.0"
segmentation-models-pytorch = "<0.2.0"
albumentations = "^1.2.1"
torch = "<=1.8.0"
torchvision = "<=0.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
